% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getIterSymbols}
\alias{getIterSymbols}
\title{Get Names of Iteration Blocks}
\usage{
getIterSymbols(sourceCode)
}
\arguments{
\item{sourceCode}{character vector with the templated strings}
}
\value{
character vector with the names of the iteration blocks
}
\description{
Get Names of Iteration Blocks
}
\details{
Iteration blocks have the following form: \cr\cr
\code{//R# <symbol name>[[start]]} \cr
\code{some text with [[R#<template variable>]]} \cr
\code{//R# <symbol name>[[end]]} \cr\cr
The name of the iteration block is given by \code{<symbol name>}.
Blocks may contain template variables of the form [[R#<template variable>]] which will be replaced during template transformation.
Distinct iteration blocks can share the same name.
}
\examples{
sourceCode <- c(
"//R# name1[[start]]","double [[R#fun1Name]](double x);","//R# name1[[end]]",
"//R# name2[[start]]","double [[R#fun2Name]](double y);","//R# name2[[end]]")

getIterSymbols(sourceCode)

}
