% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getVarSymbols}
\alias{getVarSymbols}
\title{Get Template Variables}
\usage{
getVarSymbols(iterSymbol, sourceCode)
}
\arguments{
\item{iterSymbol}{name of the iteration block}

\item{sourceCode}{character vector with templates}
}
\value{
character vector with the names of the template variables
}
\description{
Get the template variables occuring in iteration blocks with a specific name.
}
\examples{
sourceCode <- c(
"//R# name1[[start]]","double [[R#fun1Name]](double x);","//R# name1[[end]]",
"//R# name2[[start]]","double [[R#fun2Name]](double [[R#argName]]);","//R# name2[[end]]")

getVarSymbols("name1",sourceCode)
getVarSymbols("name2",sourceCode)

}
